resolver as separate module
